Add Isochrone.export_for_starfish method
This method writes the isochrone in a format useful for the StarFISH
analysis code. That is, columns of initial mass and magnitudes in each
available filter output in a no-header ascii file.

The file is also automatically named in the star fish formation of
`zNNNN_tt.tt` where NNNN is the fractional part of the metallicity
and `tt.tt` is log(age).

To support this, we add `z_code()` and `age_code` properties to
`Isochrone` to generate these. Also add a `filter_names` property to
generate an (ordered) list of filter names in the isochrone.